# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Al Bayaan AI Quran API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  clerkId: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  onboardingComplete: boolean;
  /** en | ar | so */
  language: string;
  /** quran_reading | hifdh | tajweed | tafsir | arabic */
  learningGoals: string[];
  /** beginner | intermediate | advanced */
  level: string;
  /** child | teen | adult | senior */
  ageGroup: string;
  dailyGoalMinutes: number;
  preferredQari: string;
  /** male | female | any */
  teacherPreference: string;
  xp: number;
  streakDays: number;
  createdAt: string;
}

export interface ProfileUpdate {
  displayName?: string;
  avatarUrl?: string;
  language?: string;
  learningGoals?: string[];
  level?: string;
  ageGroup?: string;
  dailyGoalMinutes?: number;
  preferredQari?: string;
  teacherPreference?: string;
}

export interface OnboardingInput {
  displayName: string;
  learningGoals: string[];
  level: string;
  ageGroup: string;
  dailyGoalMinutes: number;
  preferredQari: string;
  language: string;
  teacherPreference: string;
}

export interface Surah {
  number: number;
  /** Transliteration e.g. Al-Fatihah */
  name: string;
  nameArabic: string;
  nameTranslation: string;
  ayahCount: number;
  /** Meccan | Medinan */
  revelationType: string;
  juz: number;
  hasProgress?: boolean;
}

export interface Ayah {
  /** Global ayah number 1-6236 */
  number: number;
  numberInSurah: number;
  /** Arabic text */
  text: string;
  surahNumber: number;
  /** @nullable */
  page?: number | null;
  /** @nullable */
  juz?: number | null;
  /** @nullable */
  audioUrl?: string | null;
}

export interface AyahDetail {
  ayah: Ayah;
  /** English translation */
  translation: string;
  /** @nullable */
  transliteration: string | null;
  /** @nullable */
  tafsir?: string | null;
}

export interface SurahProgress {
  surahId: number;
  surahName: string;
  completedAyahs: number;
  totalAyahs: number;
  /** @nullable */
  lastStudied: string | null;
  /** @nullable */
  averageScore?: number | null;
}

export interface Progress {
  totalAyahsRecited: number;
  totalSurahsStarted: number;
  totalSurahsCompleted: number;
  totalRecordingMinutes: number;
  /** 0-100 */
  tajweedScore: number;
  /** 0-100 */
  accuracyScore: number;
  surahsProgress: SurahProgress[];
}

export interface SurahProgressUpdate {
  completedAyahs: number;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  xp: number;
  /** Numeric level 1-100 */
  level: number;
  xpToNextLevel: number;
  dailyGoalMinutes: number;
  dailyGoalCompletedToday: boolean;
  minutesStudiedToday?: number;
}

export interface AIFeedback {
  /** 0-100 */
  pronunciationScore: number;
  /** 0-100 */
  fluencyScore: number;
  /** 0-100 */
  tajweedScore: number;
  /** 0-100 */
  accuracyScore: number;
  /** 0-100 */
  confidenceScore: number;
  /** 0-100 */
  overallScore: number;
  correctWords: string[];
  incorrectWords: string[];
  missingWords: string[];
  suggestions: string[];
  /** @nullable */
  transcribedText?: string | null;
}

export interface Recording {
  id: number;
  userId: string;
  surahId: number;
  ayahId: number;
  ayahNumber: number;
  /** @nullable */
  audioUrl?: string | null;
  durationSeconds: number;
  createdAt: string;
  feedback?: AIFeedback;
}

export interface RecordingInput {
  surahId: number;
  ayahId: number;
  ayahNumber: number;
  /** The reference Arabic text to compare against */
  ayahText: string;
  /**
     * Base64 encoded audio for AI analysis
     * @nullable
     */
  audioBase64?: string | null;
  durationSeconds: number;
}

export interface Bookmark {
  id: number;
  userId: string;
  surahId: number;
  surahName?: string;
  ayahNumber: number;
  ayahText?: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface BookmarkInput {
  surahId: number;
  surahName: string;
  ayahNumber: number;
  ayahText: string;
  note?: string;
}

export interface Achievement {
  id: number;
  slug: string;
  title: string;
  description: string;
  /** star | flame | book | mic | trophy | medal | crown | heart | moon | sun */
  iconType: string;
  xpReward: number;
  isUnlocked: boolean;
  /** @nullable */
  unlockedAt: string | null;
  /**
     * 0-100 for locked achievements
     * @nullable
     */
  progress?: number | null;
}

export interface QuickStats {
  avgAccuracy: number;
  avgTajweed: number;
  totalRecordings: number;
  totalBookmarks: number;
  surahsInProgress: number;
  surahsCompleted: number;
}

export interface DashboardSummary {
  profile: Profile;
  streak: StreakInfo;
  todayMinutes: number;
  weeklyXp: number;
  totalAyahs: number;
  recentRecordings: Recording[];
  recentAchievements: Achievement[];
  quickStats: QuickStats;
}

export interface DayActivity {
  date: string;
  xp: number;
  minutes: number;
  recordings: number;
}

export interface WeeklyReport {
  days: DayActivity[];
  totalXp: number;
  totalMinutes: number;
  totalRecordings: number;
  avgScore: number;
  bestDay: string;
  /** Percentage change vs previous week */
  improvement: number;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  xp: number;
  streakDays: number;
  surahsCompleted: number;
  isCurrentUser?: boolean;
}

export type GetLeaderboardParams = {
period?: GetLeaderboardPeriod;
};

export type GetLeaderboardPeriod = typeof GetLeaderboardPeriod[keyof typeof GetLeaderboardPeriod];


export const GetLeaderboardPeriod = {
  weekly: 'weekly',
  monthly: 'monthly',
  alltime: 'alltime',
} as const;