# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al Bayaan AI Quran API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get current user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profile/onboarding:
    post:
      operationId: completeOnboarding
      tags: [profile]
      summary: Complete onboarding flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarding complete, profile returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # ── SURAHS ───────────────────────────────────────────────────────────────
...
  /progress/surah/{surahId}:
    get:
      operationId: getSurahProgress
      tags: [progress]
      summary: Get progress for a specific surah
      parameters:
        - in: path
          name: surahId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Surah progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurahProgress"
    patch:
      operationId: updateSurahProgress
      tags: [progress]
      summary: Update progress for a surah ayah
      parameters:
        - in: path
          name: surahId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurahProgressUpdate"
      responses:
        "200":
          description: Updated surah progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurahProgress"

  # ── RECORDINGS ────────────────────────────────────────────────────────────
  /recordings:
    get:
      operationId: listRecordings
      tags: [recordings]
      summary: List user's recordings
      responses:
        "200":
          description: Recordings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"
    post:
      operationId: createRecording
      tags: [recordings]
      summary: Submit a recording for AI analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordingInput"
      responses:
        "201":
          description: Recording created with AI feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
  /recordings/{id}:
    get:
      operationId: getRecording
      tags: [recordings]
      summary: Get recording with AI feedback
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recording detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"

  # ── BOOKMARKS ─────────────────────────────────────────────────────────────
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: List user bookmarks
      responses:
        "200":
          description: Bookmarks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Create a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmark created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{id}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Delete a bookmark
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ACHIEVEMENTS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, clerkId, displayName, onboardingComplete, language, learningGoals, level, ageGroup, dailyGoalMinutes, preferredQari, teacherPreference, xp, streakDays, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        onboardingComplete:
          type: boolean
        language:
          type: string
          description: "en | ar | so"
        learningGoals:
          type: array
          items:
            type: string
          description: "quran_reading | hifdh | tajweed | tafsir | arabic"
        level:
          type: string
          description: "beginner | intermediate | advanced"
        ageGroup:
          type: string
          description: "child | teen | adult | senior"
        dailyGoalMinutes:
          type: integer
        preferredQari:
          type: string
        teacherPreference:
          type: string
          description: "male | female | any"
        xp:
          type: integer
        streakDays:
          type: integer
        createdAt:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        avatarUrl:
          type: string
        language:
          type: string
        learningGoals:
          type: array
          items:
            type: string
        level:
          type: string
        ageGroup:
          type: string
        dailyGoalMinutes:
          type: integer
        preferredQari:
          type: string
        teacherPreference:
          type: string
    OnboardingInput:
      type: object
      required: [displayName, learningGoals, level, ageGroup, dailyGoalMinutes, preferredQari, language, teacherPreference]
      properties:
        displayName:
          type: string
        learningGoals:
          type: array
          items:
            type: string
        level:
          type: string
        ageGroup:
          type: string
        dailyGoalMinutes:
          type: integer
        preferredQari:
          type: string
        language:
          type: string
        teacherPreference:
          type: string
...
    SurahProgress:
      type: object
      required: [surahId, surahName, completedAyahs, totalAyahs, lastStudied]
      properties:
        surahId:
          type: integer
        surahName:
          type: string
        completedAyahs:
          type: integer
        totalAyahs:
          type: integer
        lastStudied:
          type: ["string", "null"]
          format: date-time
        averageScore:
          type: ["integer", "null"]
    SurahProgressUpdate:
      type: object
      required: [completedAyahs]
      properties:
        completedAyahs:
          type: integer
...
    Recording:
      type: object
      required: [id, userId, surahId, ayahId, ayahNumber, createdAt, durationSeconds]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        ayahId:
          type: integer
        ayahNumber:
          type: integer
        audioUrl:
          type: ["string", "null"]
        durationSeconds:
          type: integer
        createdAt:
          type: string
          format: date-time
        feedback:
          $ref: "#/components/schemas/AIFeedback"
    AIFeedback:
      type: object
      required: [pronunciationScore, fluencyScore, tajweedScore, accuracyScore, confidenceScore, overallScore, correctWords, incorrectWords, missingWords, suggestions]
      properties:
        pronunciationScore:
          type: integer
          description: 0-100
        fluencyScore:
          type: integer
          description: 0-100
        tajweedScore:
          type: integer
          description: 0-100
        accuracyScore:
          type: integer
          description: 0-100
        confidenceScore:
          type: integer
          description: 0-100
        overallScore:
          type: integer
          description: 0-100
        correctWords:
          type: array
          items:
            type: string
        incorrectWords:
          type: array
          items:
            type: string
        missingWords:
          type: array
          items:
            type: string
        suggestions:
          type: array
          items:
            type: string
        transcribedText:
          type: ["string", "null"]
    RecordingInput:
      type: object
      required: [surahId, ayahId, ayahNumber, ayahText, durationSeconds]
      properties:
        surahId:
          type: integer
        ayahId:
          type: integer
        ayahNumber:
          type: integer
        ayahText:
          type: string
          description: The reference Arabic text to compare against
        audioBase64:
          type: ["string", "null"]
          description: Base64 encoded audio for AI analysis
        durationSeconds:
          type: integer
    Bookmark:
      type: object
      required: [id, userId, surahId, ayahNumber, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        surahId:
          type: integer
        surahName:
          type: string
        ayahNumber:
          type: integer
        ayahText:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    BookmarkInput:
      type: object
      required: [surahId, ayahNumber, ayahText, surahName]
      properties:
        surahId:
          type: integer
        surahName:
          type: string
        ayahNumber:
          type: integer
        ayahText:
          type: string
        note:
          type: string
...