# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Al Bayaan AI Quran API
 * OpenAPI spec version: 0.1.0
 */

export interface Bookmark {
  id: number;
  userId: string;
  surahId: number;
  surahName?: string;
  ayahNumber: number;
  ayahText?: string;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}