# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Al Bayaan AI Quran API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  clerkId: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  onboardingComplete: boolean;
  /** en | ar | so */
  language: string;
  /** quran_reading | hifdh | tajweed | tafsir | arabic */
  learningGoals: string[];
  /** beginner | intermediate | advanced */
  level: string;
  /** child | teen | adult | senior */
  ageGroup: string;
  dailyGoalMinutes: number;
  preferredQari: string;
  /** male | female | any */
  teacherPreference: string;
  xp: number;
  streakDays: number;
  createdAt: Date;
}