# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Al Bayaan AI Quran API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileUpdate {
  displayName?: string;
  avatarUrl?: string;
  language?: string;
  learningGoals?: string[];
  level?: string;
  ageGroup?: string;
  dailyGoalMinutes?: number;
  preferredQari?: string;
  teacherPreference?: string;
}